fix: update pre-render with to work with new razzle export
