+++ examples/with-prerender/package.json
@@ -5,8 +5,9 @@
   "scripts": {
     "start": "razzle start",
     "build": "razzle build",
-    "prerender": "razzle prerender",
+    "export": "razzle export",
     "test": "razzle test --env=jsdom",
+    "start:static": "serve -s build/public",
     "start:prod": "NODE_ENV=production node build/server.js"
   },
   "dependencies": {
@@ -16,6 +17,7 @@
     "react": "^16.2.0",
     "react-dom": "^16.2.0",
     "react-helmet": "^6.0.0",
-    "react-router-dom": "^5.1.2"
+    "react-router-dom": "^5.1.2",
+    "serve": "^11.3.0"
   }
 }
