--- examples/with-prerender/paths.js
@@ -1,5 +1,0 @@
-const paths = () => {
-  return ['/', '/about'];
-};
-
-module.exports.default = paths;
