+++ examples/with-prerender/razzle.config.js
@@ -2,6 +2,9 @@
 
 module.exports = {
   experimental: {
-    prerender: true,
+    static_export: {
+      script_replacement: '<!-- after_static_js -->',
+      window_variable: 'AFTER_STATIC_ROUTES'
+    },
   },
 };
