+++ examples/with-prerender/src/server.js
@@ -5,21 +5,6 @@ import routes from './routes';
 const assets = require(process.env.RAZZLE_ASSETS_MANIFEST);
 const chunks = require(process.env.RAZZLE_CHUNKS_MANIFEST);
 
-export const render = async (req, res) => {
-  try {
-    const { html, data } = await renderStatic({
-      req,
-      res,
-      routes,
-      assets,
-      chunks,
-    });
-    res.json({ html, data });
-  } catch (error) {
-    res.json({ error: error.message });
-  }
-};
-
 const server = express();
 server
   .disable('x-powered-by')
