+++ packages/after.js/src/After.tsx
@@ -114,7 +114,11 @@ class Afterparty extends React.Component<AfterpartyProps, AfterpartyState> {
 
       // in ssg mode we don't call component.getInitialProps
       // instead we fetch the page-data.json file
-      const loadData = ssg ? loadStaticProps : loadInitialProps;
+      const useStaticProps = ssg
+        ? (window.AFTER_STATIC_ROUTES ? window.AFTER_STATIC_ROUTES.includes(
+              location.pathname == '' ? '/' : location.pathname) : true) : false;
+
+      const loadData = useStaticProps ? loadStaticProps : loadInitialProps;
 
       loadData(location.pathname, routes, ctx)
         .then(res => res.data)
